+++ test/test_http10.rb
@@ -2,8 +2,6 @@
 
 require_relative "helper"
 
-require "puma/puma_http11"
-
 class Http10ParserTest < PumaTest
   def test_parse_simple
     parser = Puma::HttpParser.new