+++ test/test_request_single.rb
@@ -594,9 +594,10 @@ class TestRequestHeadersInvalid < TestRequestBase
   def test_malformed_headers_no_return
     request = "GET / HTTP/1.1\r\nHost: test.com\r\nno-return: 10\nContent-Length: 11\r\n\r\nHello World"
 
-    msg = Puma::IS_JRUBY ?
-      "Invalid HTTP format, parsing fails. Bad headers\nHost: test.com\nno-return: 10\\nContent-Length: 11" :
-      "Invalid HTTP format, parsing fails. Bad headers\nHOST: test.com\nNO_RETURN: 10\\nContent-Length: 11"
+    # the C extension upcases header names in the buffer while parsing
+    msg = Puma.http_parser_engine == "c" ?
+      "Invalid HTTP format, parsing fails. Bad headers\nHOST: test.com\nNO_RETURN: 10\\nContent-Length: 11" :
+      "Invalid HTTP format, parsing fails. Bad headers\nHost: test.com\nno-return: 10\\nContent-Length: 11"
 
     assert_invalid request, msg
   end
@@ -604,9 +605,10 @@ class TestRequestHeadersInvalid < TestRequestBase
   def test_malformed_headers_no_newline
     request = "GET / HTTP/1.1\r\nHost: test.com\r\nno-newline: 10\rContent-Length: 11\r\n\r\nHello World"
 
-    msg = Puma::IS_JRUBY ?
-      "Invalid HTTP format, parsing fails. Bad headers\nHost: test.com\nno-newline: 10\rContent-Length: 11" :
-      "Invalid HTTP format, parsing fails. Bad headers\nHOST: test.com\nNO_NEWLINE: 10\rContent-Length: 11"
+    # the C extension upcases header names in the buffer while parsing
+    msg = Puma.http_parser_engine == "c" ?
+      "Invalid HTTP format, parsing fails. Bad headers\nHOST: test.com\nNO_NEWLINE: 10\rContent-Length: 11" :
+      "Invalid HTTP format, parsing fails. Bad headers\nHost: test.com\nno-newline: 10\rContent-Length: 11"
 
     assert_invalid request, msg
   end